+++ sources/packages/backend/apps/api/src/dynamic-form-configuration.module.ts
@@ -1,10 +1,15 @@
 import { Global, LoggerService, Module, OnModuleInit } from "@nestjs/common";
-import { DynamicFormConfigurationService } from "./services";
+import {
+  DynamicFormConfigurationService,
+  ProgramYearService,
+} from "./services";
 import { InjectLogger } from "@sims/utilities/logger";
+import { DynamicFormConfigurationController } from "./route-controllers";
 
 @Global()
 @Module({
-  providers: [DynamicFormConfigurationService],
+  controllers: [DynamicFormConfigurationController],
+  providers: [DynamicFormConfigurationService, ProgramYearService],
   exports: [DynamicFormConfigurationService],
 })
 export class DynamicFormConfigurationModule implements OnModuleInit {
